projects/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
click==7.1.2
ecdsa==0.16.1
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pycryptodome==3.3.1
pyflakes==2.2.0
python-dotenv==0.15.0
python-jose-cryptodome==1.3.2
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==0.15.5
wrapt==1.11.1
